add env code

+++ devops-common/devops-common-entity/src/main/java/com/hfwas/devops/entity/DevopsEnv.java
@@ -16,14 +16,17 @@ public class DevopsEnv {
     private String code;
     private String name;
     private String description;
+    /**
+     * 环境类型，1、dev，2、uat，3、pre、4、prod
+     */
     private Integer type;
 
     @TableField("project_id")
     private Integer projectId;
     @TableField(value = "create_by", fill = FieldFill.INSERT)
-    private Integer createBy;
+    private Long createBy;
     @TableField(value = "update_by", fill = FieldFill.INSERT_UPDATE)
-    private Integer updateBy;
+    private Long updateBy;
     @TableField(value = "create_time", fill = FieldFill.INSERT)
     private LocalDateTime createTime;
     @TableField(value = "update_time", fill = FieldFill.INSERT_UPDATE)
